Fix/mqtt resub

+++ src/controllers/messages.ts
@@ -1,4 +1,11 @@
-import { models, Message, Chat, ContactRecord } from '../models'
+import {
+  models,
+  Message,
+  Chat,
+  ContactRecord,
+  MessageRecord,
+  ChatRecord,
+} from '../models'
 import { Op, FindOptions } from 'sequelize'
 import { indexBy } from 'underscore'
 import {
@@ -23,6 +30,7 @@ import { ChatPlusMembers } from '../network/send'
 import { getCacheMsg } from '../utils/tribes'
 import { CronJob } from 'cron'
 import { loadConfig } from '../utils/config'
+import { onReceiveReversal } from '../utils/reversal'
 
 interface ExtentedMessage extends Message {
   chat_id?: number
@@ -32,6 +40,19 @@ interface ExtentedMessage extends Message {
 const jobs = {}
 
 const config = loadConfig()
+
+export const getMessageByUuid = async (req: Req, res: Res): Promise<void> => {
+  if (!req.owner) return failure(res, 'no owner')
+  const tenant: number = req.owner.id
+  const uuid = req.params.uuid
+  if (!uuid) return failure(res, 'no uuid supplied')
+  const message: Message = (await models.Message.findOne({
+    where: { tenant, uuid },
+  })) as Message
+  success(res, {
+    message: jsonUtils.messageToJson(message),
+  })
+}
 
 // deprecated
 export const getMessages = async (req: Req, res: Res): Promise<void> => {
@@ -506,11 +527,12 @@ export const receiveMessage = async (payload: Payload): Promise<void> => {
     person,
     cached,
   } = await helpers.parseReceiveParams(payload)
-  sphinxLogger.info(`received message on tenant ${owner.id} chat ${chat.id}`)
 
   if (!owner || !sender || !chat) {
     return sphinxLogger.info('=> no group chat!')
   }
+  sphinxLogger.info(`received message on tenant ${owner.id} chat ${chat.id}`)
+
   const tenant: number = owner.id
   const text = content || ''
 
@@ -611,6 +633,17 @@ export const receiveBoost = async (payload: Payload): Promise<void> => {
   let date = new Date()
   date.setMilliseconds(0)
   if (date_string) date = new Date(date_string)
+
+  if (payload.error_message) {
+    return await onReceiveReversal({
+      tenant,
+      type: 'boost',
+      errorMsg: payload.error_message,
+      msgUuid: payload.message.uuid,
+      chat,
+      sender,
+    })
+  }
 
   const msg: { [k: string]: string | number | Date } = {
     chatId: chat.id,
@@ -810,6 +843,20 @@ export const clearMessages = async (req: Req, res: Res): Promise<void> => {
   success(res, {})
 }
 
+export async function disappearingMessages(req: Req, res: Res): Promise<void> {
+  if (!req.owner) return failure(res, 'no owner')
+  const tenant: number = req.owner.id
+  try {
+    const contacts = (await models.Contact.findAll({
+      where: { tenant, isOwner: true },
+    })) as ContactRecord[]
+    await deleteMessages(contacts)
+    return success(res, 'Messages deleted successfully')
+  } catch (error) {
+    return failure(res, error)
+  }
+}
+
 export const receiveVoip = async (payload: Payload): Promise<void> => {
   sphinxLogger.info(`received Voip ${payload}`)
   const {
@@ -979,7 +1026,7 @@ async function deleteMessages(contacts: ContactRecord[]) {
       const contact: ContactRecord = contacts[i]
       const date = new Date()
       date.setDate(
-        date.getDate() - (contact.prune || parseInt(config.default_prune))
+        date.getDate() - (contact.prune || parseInt(config.default_prune || 0))
       )
       await handleMessageDelete({
         tenant: contact.tenant,
@@ -1002,9 +1049,27 @@ async function handleMessageDelete({
   date: string
 }) {
   try {
-    await models.Message.destroy({
-      where: { tenant, createdAt: { [Op.lt]: date } },
-    })
+    const chats = (await models.Chat.findAll({
+      where: { tenant, deleted: false },
+    })) as ChatRecord[]
+
+    for (let i = 0; i < chats.length; i++) {
+      const chat = chats[i]
+      const chatMessages = (await models.Message.findAll({
+        where: { chatId: chat.id, tenant },
+      })) as MessageRecord[]
+      if (chatMessages.length > 10) {
+        const tenthToLastID = chatMessages[chatMessages.length - 10]
+        await models.Message.destroy({
+          where: {
+            id: { [Op.lt]: tenthToLastID.id },
+            createdAt: { [Op.lt]: date },
+            chatId: chat.id,
+            tenant,
+          },
+        })
+      }
+    }
     sphinxLogger.info(['=> message deleted by cron job'])
   } catch (error) {
     sphinxLogger.error(['=> error deleting message by cron job', error])
